Type

Documentation

newtype FunId

Constructors

FunId Pi

type Ann = Set FunId

newtype AnnVar

Constructors

AnnVar Integer

newtype TyVar

Constructors

TyVar Integer

data Ty a

A Ty a represents a type with type annotations of type a. A TArrow t1 a b t2 is the t1 -> t2 function type, which may have been defined at program points in a, and may call functions defined at points in b. The TPair and TList constructors respectively represent pairs and lists, where the a annotation contains all locations they could have been created.

Constructors

FreeVar TyVar

TInt

TBool

TArrow (Ty a) a a (Ty a)

TPair a (Ty a) (Ty a)

TList (Ty a) a

data TyScheme a

Constructors

SType (Ty a)

Forall TyVar (TyScheme a)

type TyEnv = Map Name (TyScheme AnnVar)

data TySubst

A TySubst represents the composite substitution from type variables to types, and annotation variables to annotation variables.

Constructors

TySubst (Map TyVar (Ty AnnVar)) (Map AnnVar AnnVar)

data Constr

A Constr represents a relation between an annotation variable and either another variable, or a program point. Ni a pi expresses that pi is in a, and Super a b indicates that a contains b.

Constructors

Ni AnnVar FunId

Super AnnVar AnnVar

type Constrs = Set Constr

typeOf :: Expr -> (TyScheme Ann, Ann)

Infer the type and annotation of a given (top-level) expression.

ctaW' :: Expr -> (Ty AnnVar, AnnVar, TySubst, Constrs)

Run ctaW on an expression.

algorithm W

ctaW :: TyEnv -> Expr -> State Integer (Ty AnnVar, AnnVar, TySubst, Constrs)

ctaW implements algorithm W adapted to Call Tracking Analysis.

subtyping

subUnify :: Ty AnnVar -> Ty AnnVar -> State Integer (TySubst, Ty AnnVar, Constrs)

Unify the subtypes of the argument types.

subtype :: Ty AnnVar -> Variance -> State Integer (Ty AnnVar, Constrs)

Create a subtype of the given type.

data Variance

A Variance tracks the polarity of locations in types. For an arrow t1 -> t2 in positive position, t1 is negative, and vice versa.

Constructors

Co

Contra

op :: Variance -> Variance

variance :: Variance -> AnnVar -> AnnVar -> Constr

If a and b occur in the same location in t1 and t2 respectively, return the necessary constraint to maintain t1 <= t2. This means that if a is covariant or positive, we return a <= b, and b <= a if it is contravariant or negative.

unification

unify :: Ty AnnVar -> Ty AnnVar -> (TySubst, Ty AnnVar)

Unify two types, returning the substition and unified type.

chk :: TyVar -> Ty a -> TySubst -> TySubst

Check whether the given type variable and type are unifiable. This wrapper for chk' ensures that a unifies with FreeVar a, while this would not unify after stripping of a type constructor, which would create an infinite type.

chk' :: TyVar -> Ty a -> TySubst -> TySubst

Check whether the given type variable and type are unifiable. Note that a cannot unify with FreeVar a here, since this would indicate the overall type becomes infinite.

helpers

freshVar :: State Integer (Ty a)

Create a fresh type variable.

freshAnnVar :: State Integer AnnVar

Create a fresh annotation variable.

annUnion :: [AnnVar] -> State Integer (AnnVar, Constrs)

For as, create an annotation variable a and the constraints expressing that a contains the union of as.

annNi :: AnnVar -> Pi -> Constrs -> Constrs

annNi a pi inserts the constraint expressing that pi is in a.

genIn :: Name -> Ty AnnVar -> TyEnv -> TyEnv

Generalise and insert a variable into an environment.

cfaInstantiate :: TyScheme AnnVar -> State Integer (Ty AnnVar)

Instatiate a type scheme to a type by replacing all bound variables with fresh ones.

generalise :: TyEnv -> Ty a -> TyScheme a

Generalise a type to a type scheme by quantifying all variables which appear freely in the type and are unbound in the environment.

envFreeVars :: TyEnv -> Set TyVar

schemeFreeVars :: TyScheme a -> Set TyVar

freeVars :: Ty a -> Set TyVar

substituting

class Substitute a where

Substitute a states that for each s :: TySubst, there is a meaningful substitution function subst s :: a -> a. In particular subst should obey the identity laws

subst mempty = id

subst (TySubst (M.singleton a (FreeVar a)) mempty) = id

subst (TySubst mempty (M.singleton a a)) = id

and idempotence

subst a . subst a = subst a

(We could narrow down the description further with alpha-equivalence lemmas (e.g. injective substitutions are bijections, and preserve alpha-equivalence both ways), but this is left as an exercise to the reader.)

Methods

subst :: TySubst -> a -> a

substitute :: Ord a => a -> Map a a -> a

substitute x s is the base case for most substitutions, replacing x with its substitution if it is in s, and returning x otherwise.

constraints solving

simplify :: Constrs -> (TySubst, Constrs)

Return the substitution and simplified constraint set, formed by removing all cyclic components and substituting all variables in such components to a representative.

toGraph :: Constrs -> [(AnnVar, AnnVar, [AnnVar])]

Convert a set of constraints to the constraint graph, in which a directed edge a->b indicates a subset relation a <= b.

deloop :: Constrs -> Constrs

Remove all constraint loops of the form Super a a.

solveConstraints :: Constrs -> AnnVar -> Ann

Greedily solve a set of constraints for the least solution. Loops when the constraints contain a cyclic component.

replaceAnnVar :: (AnnVar -> Ann) -> Ty AnnVar -> Ty Ann